Improve TTY handling for host commands
If we are running the pegg command line from a terminal, then the
subprocess (which is in a different session) can't set the terminal
as it's controlling terminal. To work around this, open a PTY and
forward input to the PTY (setting the original terminal to raw mode.)

Also work around a problem in older versions of Flatpak where all commands
- even when we didn't want a terminal - tried to take ownership of the
terminal by forwarding non-TTY commands via pipes.